refactor: fix findings from ruff and vulture
- bench: 루프 변수를 닫는 클로저를 기본 인자로 묶고, 쓰이지 않는 지역 변수 제거
- application: 이미 정수인 round() 결과를 다시 int() 로 감싸던 부분 정리
- tests: zip(strict=True)

vulture 가 짚은 do_GET/do_POST 는 BaseHTTPRequestHandler 가 부르는 훅이라
오탐이다. 저장소에 린터 설정은 두지 않았고(의존성 없이 유지), uvx 로
ruff·vulture 를 한 번 돌려 확인했다.

+++ src/application.py
@@ -108,8 +108,8 @@ def preprocess_image(image):
     pixels = np.asarray(canvas, dtype=np.float32)
     y, x = np.indices(pixels.shape)
     mass = pixels.sum()
-    dy = int(round(13.5 - (y * pixels).sum() / mass))
-    dx = int(round(13.5 - (x * pixels).sum() / mass))
+    dy = round(13.5 - (y * pixels).sum() / mass)
+    dx = round(13.5 - (x * pixels).sum() / mass)
     centered = Image.new("L", (28, 28))
     centered.paste(canvas, (dx, dy))
     return np.asarray(centered, dtype=np.float32).reshape(1, 784) / 255.0
